re-organize file tree
